Update SqlServerConnection to empty string in appsettings.json
The SqlServerConnection value in the ConnectionStrings section of
the appsettings.json file was updated from an unspecified value
(indicated by a comma) to an empty string. This change explicitly
sets the SqlServerConnection to an empty value, possibly to clear
out any previous connection string or to prepare for a new value
to be set later.